+++ test/vec2/vec2_random_test.exs
@@ -1,11 +1,13 @@
 defmodule Graphmath.Vec2.RandomVec2 do
   use ExUnit.Case
 
+  @very_smol 1.0e-10
+
   @tag :vec2
   @tag :random
   test "random_circle/0 works" do
     v = Graphmath.Vec2.random_circle()
-    assert Graphmath.Vec2.length_squared(v) == 1.0
+    assert abs(1.0 - Graphmath.Vec2.length_squared(v)) < @very_smol
   end
 
   @tag :vec2