+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2021-06-14
+### Changed
+- Update to require Elixir 12.1.
+- Update credo to 1.5.6.
+- Update dialyxir to 1.1.0.
+- Update ex_doc to 0.24.2.
+- Update excoveralls to 0.14.1.
+- Update use of `:random.uniform/0` to be `:rand.uniform/0`.
+
 ## [2.4.0] - 2019-07-28
 ### Added
 - Add `Vec3.scalar_triple` product for vec3s.